Update nmap scan
I have added two different types of TCP Connect Scans. 
First one: 
-sT: Specifies a TCP Connect Scan
-sV: Enables version detection to determine the version of services running on the target
-p 1-65535: Specifies the range of ports to scan
-oN: Specifies the output format as a normal text file
target_ip_address: Specifies the IP address of the target to scan

Second One: 
-sS: Specifies a SYN Scan, which is faster than a TCP Connect Scan.
-sU: Enables a UDP Scan in addition to the SYN Scan.
-sV: Enables version detection to determine the version of services running on the target.
-A: Enables OS and version detection, script scanning, and traceroute.
-O: Enables operating system detection.
-p 1-65535: Specifies the range of ports to scan.
-T4: Sets the timing template to aggressive, which speeds up the scan but may increase the risk of detection.
-oA scan_results: Specifies the output format as three different file types (normal, XML, and grepable) with a prefix of scan_results.
target_ip_address: Specifies the IP address of the target to scan.
This scan will perform a SYN Scan and UDP Scan on all ports in the range of 1-65535, and will attempt to identify the versions of the services running on those ports, as well as the operating system of the target. It will also enable script scanning, traceroute, and aggressive timing to speed up the scan. The output will be saved in three different file types with the prefix scan_results.